#334 move geometry tests into separate files
